requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
ipykernel>=6.29.5
pandas>=2.2.0
datasets>=3.2.0
pytest>=7.0.0
